pyproject.toml
[project]
name = "birdle"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.7.2",
    "beautifulsoup4==4.12.3",
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "dj-database-url==2.1.0",
    "django==5.0.1",
    "django-htmx==1.17.3",
    "django-on-heroku==1.1.2",
    "django-stubs>=5.0.0",
    "gunicorn==21.2.0",
    "idna==3.6",
    "numpy==1.26.3",
    "packaging==23.2",
    "pandas==2.2.0",
    "psycopg2==2.9.9",
    "psycopg2-binary==2.9.9",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.1",
    "pytz==2023.3.post1",
    "requests==2.31.0",
    "six==1.16.0",
    "soupsieve==2.5",
    "sqlparse==0.4.4",
    "typing-extensions==4.9.0",
    "tzdata==2023.4",
    "urllib3==2.1.0",
    "whitenoise==6.6.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W"]

# Ignore Django-specific type errors that require django-stubs
# The 'objects' manager and 'DoesNotExist' exception are added dynamically by Django
[tool.ty.rules]

